requirements.txt
cffi==1.11.5
click==6.7
colorama==0.3.9
colorlog==3.1.4
colormap==1.0.1
easydev==0.9.37
Flask==1.0.2
gevent==1.3.6
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pexpect==4.6.0
Pillow==5.2.0
ptyprocess==0.6.0
pycparser==2.18
qrcode==6.0
six==1.11.0
Werkzeug==0.14.1
